# server/public/swagger.json
{
  "openapi":"3.1.0","info":{"title":"FormFlow Studio API","version":"0.5.0","description":"FormFlow projects, data, workflows, tasks, governance and integrations API"},
  "servers":[{"url":"/api"}],
  "components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"Error":{"type":"object","properties":{"error":{"type":"string"}}},"Task":{"type":"object","properties":{"id":{"type":"string"},"state":{"enum":["queued","running","completed","failed","cancelled"]},"progress":{"type":"number"}}}}},
  "paths":{
...
    "/projects":{"get":{"summary":"List accessible projects","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Projects"}}},"post":{"summary":"Create project","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Created"}}}},
...
  }
}